Remove vavr dependecy from Bulkhead and RateLimiter (#1141)
