+++ resilience4j-feign/src/main/java/io/github/resilience4j/feign/FeignDecorators.java
@@ -19,9 +19,11 @@ package io.github.resilience4j.feign;
 import feign.InvocationHandlerFactory.MethodHandler;
 import feign.Target;
 import io.github.resilience4j.bulkhead.Bulkhead;
+import io.github.resilience4j.bulkhead.VavrBulkhead;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.circuitbreaker.VavrCircuitBreaker;
 import io.github.resilience4j.ratelimiter.RateLimiter;
+import io.github.resilience4j.ratelimiter.VavrRateLimiter;
 import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.VavrRetry;
 import io.vavr.CheckedFunction1;
@@ -113,7 +115,7 @@ public class FeignDecorators implements FeignDecorator {
          * @return the builder
          */
         public Builder withRateLimiter(RateLimiter rateLimiter) {
-            addFeignDecorator(fn -> RateLimiter.decorateCheckedFunction(rateLimiter, fn));
+            addFeignDecorator(fn -> VavrRateLimiter.decorateCheckedFunction(rateLimiter, fn));
             return this;
         }
 
@@ -215,7 +217,7 @@ public class FeignDecorators implements FeignDecorator {
          * @return the builder
          */
         public Builder withBulkhead(Bulkhead bulkhead) {
-            addFeignDecorator(fn -> Bulkhead.decorateCheckedFunction(bulkhead, fn));
+            addFeignDecorator(fn -> VavrBulkhead.decorateCheckedFunction(bulkhead, fn));
             return this;
         }
