+++ resilience4j-ratpack/src/main/java/io/github/resilience4j/ratpack/ratelimiter/RateLimiterMethodInterceptor.java
@@ -19,6 +19,7 @@ package io.github.resilience4j.ratpack.ratelimiter;
 import com.google.inject.Inject;
 import io.github.resilience4j.core.lang.Nullable;
 import io.github.resilience4j.ratelimiter.RateLimiterRegistry;
+import io.github.resilience4j.ratelimiter.VavrRateLimiter;
 import io.github.resilience4j.ratelimiter.annotation.RateLimiter;
 import io.github.resilience4j.ratpack.internal.AbstractMethodInterceptor;
 import io.github.resilience4j.ratpack.recovery.DefaultRecoveryFunction;
@@ -129,8 +130,7 @@ public class RateLimiterMethodInterceptor extends AbstractMethodInterceptor {
         io.github.resilience4j.ratelimiter.RateLimiter rateLimiter,
         RecoveryFunction<?> recoveryFunction) throws Throwable {
         try {
-            return io.github.resilience4j.ratelimiter.RateLimiter
-                .decorateCheckedSupplier(rateLimiter, invocation::proceed).apply();
+            return VavrRateLimiter.decorateCheckedSupplier(rateLimiter, invocation::proceed).apply();
         } catch (Throwable t) {
             return recoveryFunction.apply(t);
         }
