+++ resilience4j-spring/src/main/java/io/github/resilience4j/ratelimiter/configure/RateLimiterAspect.java
@@ -19,6 +19,7 @@ import io.github.resilience4j.core.lang.Nullable;
 import io.github.resilience4j.fallback.FallbackExecutor;
 import io.github.resilience4j.ratelimiter.RateLimiterConfig;
 import io.github.resilience4j.ratelimiter.RateLimiterRegistry;
+import io.github.resilience4j.ratelimiter.VavrRateLimiter;
 import io.github.resilience4j.ratelimiter.annotation.RateLimiter;
 import io.github.resilience4j.spelresolver.SpelResolver;
 import io.github.resilience4j.utils.AnnotationExtractor;
@@ -167,7 +168,7 @@ public class RateLimiterAspect implements Ordered {
     private Object handleJoinPoint(ProceedingJoinPoint proceedingJoinPoint,
         io.github.resilience4j.ratelimiter.RateLimiter rateLimiter)
         throws Throwable {
-        return rateLimiter.executeCheckedSupplier(proceedingJoinPoint::proceed);
+        return VavrRateLimiter.executeCheckedSupplier(rateLimiter, proceedingJoinPoint::proceed);
     }
 
     /**
